Refactor Coursify Studio UI and optimize performance/engagement
Summary of changes:
- Phase 1: Modularized the Coursify Studio UI by extracting state and logic into a new `CoursifyStudioContext`. Created domain-specific components: `AuthoringHeader`, `PlanningWorkspace`, and `ModuleSectionList`. This significantly reduced the complexity and size of the main studio page.
- Phase 2: Improved Performance & SEO.
    - Converted the public course listing page to a Server Component for faster initial load.
    - Created `CourseListClient` to handle interactive filtering and search.
    - Enabled Incremental Static Regeneration (ISR) for course detail and section pages.
    - Integrated published courses into the `sitemap.xml` route.
    - Optimized image delivery using the Next.js `Image` component in `CourseCard` and `CourseOverview`.
- Phase 3: Enhanced Learner Engagement.
    - Implemented local-first progress tracking by persisting "visited" sections in `localStorage`.
    - Added a "Study Streak" counter using a new `useStudyStreak` hook, displayed in the reader header.
    - Persisted quiz progress and scores in `localStorage` within `QuizPlayer`.

Note: Phase 4 (Advanced Authoring with Drag-and-Drop) and the AI Tutor contextual suggestions were not implemented due to time/turn constraints.
